:recycle: refactor: Removendo tratamento de erros dos repositories

+++ backend/app/repositories/user_repository.py
@@ -1,12 +1,15 @@
 from sqlmodel import Session, select
 from typing import List, Optional
 from app.models import User
+from app.exceptions import NotFoundError, BusinessRuleError
 
 class UserRepository:
     def __init__(self, session: Session):
         self.session = session
 
     def create(self, name: str) -> User:
+        if not name.strip():
+            raise BusinessRuleError("USER_NAME_REQUIRED")
         user = User(name=name)
         self.session.add(user)
         self.session.commit()
@@ -14,7 +17,10 @@ class UserRepository:
         return user
 
     def get_by_id(self, user_id: int) -> Optional[User]:
-        return self.session.get(User, user_id)
+        user =  self.session.get(User, user_id)
+        if not user:
+            raise NotFoundError("USER_NOT_FOUND")
+        return user
 
     def get_all(self) -> List[User]:
         return self.session.exec(select(User)).all()
@@ -29,7 +35,10 @@ class UserRepository:
         self.session.refresh(user)
         return user
 
-    def delete(self, user: User):
+    def delete(self, user_id: int):
+        user = self.get_by_id(user_id)
+        if not user:
+            raise NotFoundError("USER_NOT_FOUND")
         self.session.delete(user)
         self.session.commit()
         return
